Experimental biquad RC fitler instead of RC interpolation

+++ src/main/common/filter.h
@@ -61,7 +61,7 @@ void pt1FilterSetTimeConstant(pt1Filter_t *filter, float tau);
 float pt1FilterGetLastOutput(pt1Filter_t *filter);
 float pt1FilterApply(pt1Filter_t *filter, float input);
 float pt1FilterApply3(pt1Filter_t *filter, float input, float dT);
-float pt1FilterApply4(pt1Filter_t *filter, float input, uint16_t f_cut, float dt);
+float pt1FilterApply4(pt1Filter_t *filter, float input, float f_cut, float dt);
 void pt1FilterReset(pt1Filter_t *filter, float input);
 
 void rateLimitFilterInit(rateLimitFilter_t *filter);
